feat: add asap trigger mode
