+++ .mockery.yaml
@@ -126,6 +126,10 @@ packages:
     config:
       dir: "{{ .InterfaceDir }}/mocks"
       all: true
+  github.com/agglayer/aggkit/aggsender/trigger/types:
+    config:
+      dir: "{{ .InterfaceDir }}/mocks"
+      all: true
   github.com/agglayer/aggkit/l2gersync:
     config:
       dir: "{{ .InterfaceDir }}/mocks"
