+++ aggsender/trigger/epoch_notifier_per_block_test.go
@@ -1,4 +1,4 @@
-package aggsender
+package trigger
 
 import (
 	"context"
